Precompiled headers, creation and streams connection ============================================================
System Information:
Engine Version: Atom Ver 0.12.1.20181013
Library Version:  Ver 0.13.2.20181013

OS/HW Interface:
Win32 Ver 0.10.8.20181013
GNU Ver 0.10.3.20181013

Hosts:
RX Interactive Console Ver 0.13.1.20181013
Win32 Console Ver 0.10.5.20181013
GNU Console Ver 0.11.4.20181013

Compilers:
MSC 14.15.26730
GCC 7.3.0

OSes:
Linux Ver:4.15.0-36-generic [x86_64]
Windows Workstation 10.0.17134 [x64]

+++ os_itf/windows/rx_win.c
@@ -4,7 +4,7 @@
 *
 *  os_itf\windows\rx_win.c
 *
-*  Copyright (c) 2017 Dusan Ciric
+*  Copyright (c) 2018 Dusan Ciric
 *
 *  
 *  This file is part of rx-platform
@@ -26,7 +26,6 @@
 ****************************************************************************/
 
 
-#include "stdafx.h"
 
 
 // rx_win
@@ -935,8 +934,8 @@ void rx_collect_memory_info(size_t* total, size_t* free, size_t* process)
 
 	GlobalMemoryStatusEx(&statex);
 
-	*total = statex.ullTotalPhys;
-	*free = statex.ullAvailPhys;
+	*total = (size_t)statex.ullTotalPhys;
+	*free = (size_t)statex.ullAvailPhys;
 }
 
 ///////////////////////////////////////////////////////////////////////////////////////////////////
